chore: Refactor book components out of `/common`
